Create Bob's Budget Code
This program will prompt the user for information in regards to 3 book collections in which a customer "Bob" is interested in buying.  The number of volumes and the price of each volume (both as integers) will be needed for this program to run (with a space in between them).
The user will also be prompted with how much the customer "Bob" wants to spend (bobsBudget). Based upon the inputs the user provided, the code will execute and print 7 pieces of information (2 additional statements compared to collections.c) about the Bob's budget and the book 
collections he is interested in buying.